Move new game state to atoms
