+++ src/core/game.ts
@@ -1,8 +1,10 @@
 import { atom, useAtom } from "jotai";
 
-import { GameConfig, GameMode } from "components/Game/types";
+import { GameConfig, GameMode, GameSettings } from "components/Game/types";
 import { parseTarget } from "services/segments";
-import { validateWord } from "services/words";
+import { getRandomWord, validateWord } from "services/words";
+import { getRandomEquation } from "services/equation";
+import { getValidateFunction } from "services/utils";
 
 const defaultGameConfig = {
   mode: GameMode.Letters,
@@ -11,4 +13,39 @@ const defaultGameConfig = {
   validate: validateWord,
 };
 
-export const useGameConfig = () => useAtom(atom<GameConfig>(defaultGameConfig));
+const gameConfigAtom = atom<GameConfig>(defaultGameConfig);
+export const useGameConfig = () => useAtom(gameConfigAtom);
+
+const gameSettingsAtom = atom<GameSettings | null>(null);
+export const useGameSettings = () => useAtom(gameSettingsAtom);
+
+export const useNewGame = () => {
+  const [_, setGameConfig] = useGameConfig();
+  const [__, setGameSettings] = useGameSettings();
+
+  const createNewGame = (
+    settings: GameSettings,
+    shouldSaveSettings: boolean
+  ) => {
+    const { mode, targetLength, difficulty } = settings;
+
+    const startedAt = new Date().toISOString();
+
+    const targetSegments =
+      mode === GameMode.Letters
+        ? getRandomWord(targetLength)
+        : getRandomEquation(difficulty);
+
+    const validate = getValidateFunction(mode);
+
+    if (shouldSaveSettings) {
+      setGameSettings(settings);
+    } else {
+      setGameSettings(null);
+    }
+
+    setGameConfig({ startedAt, mode, targetSegments, validate });
+  };
+
+  return { createNewGame };
+};
